requirements.txt
pandas 
matplotlib 
opencv-python 
ipywidgets 
numpy 
Pillow 
tqdm 
wandb